Enable OMRChecker linting on Travis CI
OMRChecker is a clang plugin designed to look for issues in the
implementation of extensible classes. OMRChecker is already
used in the Eclipse OMR Travis CI builds to report issues around
extensible class implementation and usage. Since OpenJ9
makes use of extensible classes, this tool will help ensure
that the extensible classes are implemented correctly and avoid
potential bugs.

Issue: #729
Signed-off-by: Nazim Uddin Bhuiyan <nazimudd@ualberta.ca>