Prepare JIT NASM files to be used on x86_64 Windows
File paths have to be handled differently on Windows. This is
because NASM does not deal with how OS file paths can differ, and
simply tries searching for files to include by prepending the
include paths provided to a file until it is found. This is also
the reason behind requiring trailing forward (or backward on Windows)
slashes in the include paths.

Signed-off-by: Nazim Uddin Bhuiyan <nazimudd@ualberta.ca>