Enable sharing of some of the NASM code across both 32 and 64bit targets
Operand GPRs in code that is shared across both 32 and 64 bit
targets are prefixed with an underscore when we need to, and gets
replaced with the appropriate name using the macros defined in
jilconsts.inc.

Furthermode, the _rel macro helps us share code between 32 and 64
bit hosts where we need to use the RIP-relative address on 64bit mode
for accessing data. By default, NASM generates absolute addresses
in 64bit mode and hence we need to use the rel keyword to get the
RIP-relative address.

Signed-off-by: Nazim Uddin Bhuiyan <nazimudd@ualberta.ca>