Recognize Class.cast and transform calls to it into checkcast
This provides the opportunity to convert the call into a completely
typical checkcast with a compile-time constant class bound (loadaddr).
Even if the class bound can't be constant-folded, or if for some reason
we simply fail to fold it, a checkcast whose class bound is not constant
at compile time should still be at least as good as a call to Class.cast
regardless of whether the call is inlined.

When the cast fails, the innermost frame listed in the stack trace will
identify the call to Class.cast rather than an explicit throw within
Class.cast itself. Additionally, the exception message will be generated
in the same way as for checkcast bytecode instructions, which may (and
appears to currently) differ cosmetically from the message used in
Class.cast. Neither of these discrepancies should impede debugging.

Class.cast calls occur very often in compilations involving OpenJDK
MethodHandles.