Add RecognizedCallTransformer for linkToVirtual
This changeset implements a mechanism to transform
java/lang/invoke/MethodHandle.linkToVirtual methods, in cases
where the MethodHandleTransformer is unable to transform it
due to the MemberName being unknown at compile time.

linkToVirtual is a VM INL call that would result in a j2i
transition, only for the interpreter loop to figure out the
target method and call that method, making an i2j transtion if
that method is already compiled. While it may not be possible
to know the target method of a linkToVirtual call at compile
time, it is possible to construct trees to figure that out at
runtime. The mechanism uses the JITHelper method dispatchVirtual
to represent a call into the virtual function table entry, which
may either be a thunk or a compiled method address. In either
case, the dispatch can be done the same way.

There are some cases where the vtable slot index is 0 for
private virtual methods, with no vtable entries for them. The
target method for that case is obtained from the MemberName
object, similar to linkToStatic/Special.

Changes included in this commit include:
* Add JITHelpers.dispatchVirtual as recognized method
* Add symbol for MemberName.vmIndex
* Add Symref table API to create/lookup symbol for
  J9JNIMethodID.vTableIndex
* Add implementation for the linkToVirtual transformer in
  RecognizedCallTransformer
* Update backend code (PPC, x86 family, S390) to handle
  JITHelpers.dispatchVirtual similar to
  ComputedCalls.dispatchVirtual

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>