Inline FE supressInliningRecognizedInitialCallee into inliner policy
There are no other callers. This will make it easier to refactor to
consolidate the methods to suppress, and to suppress them (in most
cases) regardless of whether inliner has a call node.

This commit's diff is best viewed with a tool that detects and shows
moved code, e.g. git show --color-moved.