Group recognized VarHandle operator methods for simplicity
This changeset categorizes VH operator methods such that all the
information relevant for handling them are available while also
significantly shortening the list of recognized methods related to
VH operations.

Added checks that ensure that non VH operator methods are not
recognized as VH operator methods based on whether the leading
parameter is VarHandle type and whether the method is static.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>