Enable peeking ILGen for param loads using NeedsPeekingHeuristic
NPH keeps track of loads of parameters when the caller method has
prex arg info, which is then used when we come across one of the
invoke* bytecodes. If the param load is used as arg for the call,
then we would want to propagate the prex arg info down the
callee, for which peeking ILGen is necessary. This commit also
terminates NPH once _needsPeeking is set as true, as any further
analysis would not change the peeking outcome. Peeking ILGen due
to param loads can be disabled using the env option TR_disableNPH.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>